// src/schemas/profile.ts
const AddressTypesSchema = z.union([
  z.literal("PLANET"),
  z.literal("CONTINENT"),
  z.literal("REGION"),
  z.literal("COUNTRY"),
  z.literal("STATE"),
  z.literal("CITY"),
  z.literal("TOWN"),
  z.literal("DISTRICT"),
  z.literal("LOCALITY"),
  z.literal("PINCODE"),
  z.literal("STREET"),
  z.literal("BUILDING"),
  z.literal("LANDMARK"),
]);

const DegreeSchema = z.union([
  z.literal("SECONDARY"),
  z.literal("HIGHER_SECONDARY"),
  z.literal("BACHELORS"),
  z.literal("MASTERS"),
  z.literal("DOCTORATE"),
]);

const RatingTypeSchema = z.union([
  z.literal("GPA"),
  z.literal("CGPA"),
  z.literal("PERCENT"),
  z.literal("PERCENTILE"),
  z.literal("NUMBER"),
]);

const RatingSchema = z.object({
  rating: z.number(),
  maxRating: z.number(),
  ratingType: RatingTypeSchema,
});
// ...
export const BasicInfoSchema = z.object({
  name: z.string(),
  job: z.string(),
  aboutMe: z.string(),
});

export const AddressSchema = z.object({
  type: AddressTypesSchema,
  name: z.string(),
});

export const PhoneSchema = z.object({
  country: CountrySchema,
  number: z.number(),
});

export const EmailSchema = z.string();

export const HobbySchema = z.string();

export const LinkSchema = z.object({
  label: z.string(),
  url: z.string(),
});

export const SkillSchema = z.object({
  name: z.string(),
  rating: RatingSchema,
});

export const LanguageSchema = z.object({
  name: z.string(),
  rating: RatingSchema,
});

export const WorkExperienceSchema = z.object({
  employer: z.string(),
  position: z.string(),
  start: z.coerce.date(),
  end: z.coerce.date().optional(),
  location: z.string(),
  remote: z.boolean(),
  description: z.union([z.string(), z.array(z.string())]),
  technologies: z.array(z.string()).optional(),
  clients: z.array(z.string()),
});

export const EducationSchema = z.object({
  school: z.string(),
  degree: DegreeSchema,
  degreeName: z.string(),
  board: z.string().optional(),
  specialization: z.string().optional(),
  start: z.coerce.date(),
  end: z.coerce.date().optional(),
  location: z.string(),
  rating: RatingSchema.optional(),
  graduated: z.boolean(),
});

export const ProjectsSchema = z.object({
  name: z.string(),
  technologies: z.array(z.string()),
  description: z.union([z.string(), z.array(z.string())]),
  sourceCode: LinkSchema.optional(),
  demo: LinkSchema.optional(),
});

export const ProfileSchema = z.object({
  basicInfo: BasicInfoSchema,
  address: z.array(AddressSchema),
  phone: PhoneSchema,
  email: EmailSchema,
  socials: z.array(LinkSchema),
  skills: z.array(SkillSchema),
  hobbies: z.array(HobbySchema),
  languages: z.array(LanguageSchema),
  experience: z.array(WorkExperienceSchema),
  education: z.array(EducationSchema),
  projects: z.array(ProjectsSchema),
});

export type TProfile = z.infer<typeof ProfileSchema>;